+++ test/libsolidity/semanticTests/structs/memory_struct_named_constructor.sol
@@ -1,0 +1,18 @@
+pragma experimental ABIEncoderV2;
+
+contract C {
+    struct S {
+        uint256 a;
+        bool x;
+    }
+
+    function s() public returns(S memory)
+    {
+        return S({x: true, a: 8});
+    }
+}
+
+// ====
+// compileViaYul: also
+// ----
+// s() -> 8, true